Widget Skeleton
Got the core skeleton of the widget completed.  This plugin can be
activated and a blank widget will show up in the widget area.  Next I
need to add the "meat" of the widget.
I added a new class (WP_Random_Image_Widget_Plugin) that handles all the
setup and hooking for the widget.